Atualização II 09/03/2021

+++ cronjob/ranking_accounts.php
@@ -55,11 +55,13 @@ foreach ($users->result_array() as $user) {
 $rank	= 1;
 $users	= Recordset::query('SELECT `id`,`score`,`level` FROM `ranking_accounts` ORDER BY `score` DESC, `level` DESC');
 foreach($users->result_array() as $user) {
-    /*if ($user->score <= 0)
+    if ($user->score <= 0)
         $user->delete();
-    else {*/
+    else {
         Recordset::update('ranking_accounts', [
             'position_general'  => $rank++
         ], ['id' => $user['id']]);
-    // }
+    }
 }
+
+echo '[Ranking Accounts] Cron executada com sucesso!';
